+++ src/wtforms/fields/__init__.py
@@ -1,4 +1,6 @@
+from wtforms.fields.choices import Choice
 from wtforms.fields.choices import RadioField
+from wtforms.fields.choices import SelectChoice
 from wtforms.fields.choices import SelectField
 from wtforms.fields.choices import SelectFieldBase
 from wtforms.fields.choices import SelectMultipleField
@@ -38,6 +40,8 @@ __all__ = [
     "Field",
     "Flags",
     "Label",
+    "Choice",
+    "SelectChoice",
     "SelectField",
     "SelectMultipleField",
     "SelectFieldBase",